Added button to explore traces from mega select
