+++ public/app/plugins/datasource/tempo/utils.ts
@@ -1,5 +1,7 @@
-import { DataSourceApi } from '@grafana/data';
+import { DataSourceApi, FieldType } from '@grafana/data';
 import { getDataSourceSrv } from '@grafana/runtime';
+
+import { AdHocVariableFilter, TraceqlFilter, TraceqlSearchScope } from './dataquery.gen';
 
 export const getErrorMessage = (message: string | undefined, prefix?: string) => {
   const err = message ? ` (${message})` : '';
@@ -19,4 +21,30 @@ export async function getDS(uid?: string): Promise<DataSourceApi | undefined> {
     console.error('Failed to load data source', error);
     return undefined;
   }
+}
+
+export function megaToFilters(megaFilters?: AdHocVariableFilter[], megaSpan?: string) {
+  const filters: TraceqlFilter[] = [];
+  if (megaSpan) {
+    filters.push({
+      id: 'mega-span',
+      tag: 'http.url',
+      operator: '=',
+      scope: TraceqlSearchScope.Span,
+      value: megaSpan,
+      valueType: FieldType.string,
+    });
+  }
+
+  filters.push(
+    ...(megaFilters?.map((f, i) => ({
+      id: f.key + i,
+      tag: f.key,
+      operator: f.operator,
+      scope: TraceqlSearchScope.Span,
+      value: f.value,
+      valueType: f.key.includes('.') ? FieldType.string : FieldType.number,
+    })) || [])
+  );
+  return filters;
 }
